1b-01.c: Warning fix.
	* gcc.dg/vmx/1b-01.c: Warning fix.
	* gcc.dg/vmx/1c-02.c: Likewise.
	* gcc.dg/vmx/gcc-bug-4.c: Likewise.

From-SVN: r110044

+++ gcc/testsuite/gcc.dg/vmx/1b-01.c
@@ -5,7 +5,7 @@ vector unsigned char a;
 static void test()
 { 
   check(sizeof(a) == 16, "sizeof(a)");
-  check(((int)&a & 15) == 0, "alignof(a)");
-  check((int)&a != 0, "&a");
+  check(((long)&a & 15) == 0, "alignof(a)");
+  check((long)&a != 0, "&a");
   check(vec_all_eq(a,((vector unsigned char){0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0})), "value(a)");
 }
