+++ gcc/testsuite/gcc.dg/vmx/gcc-bug-4.c
@@ -1,5 +1,5 @@
 /* { dg-do compile } */
 #include <altivec.h>
 extern vector unsigned int gn00111;
-int f() { return (int)&gn00111; }
+long f() { return (long)&gn00111; }
